# doc/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "summary": "获取分类详情",
        "tags": [
          "分类"
        ]
      }
    },
...
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "summary": "商品详情",
        "tags": [
          "商品（公开）"
        ]
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "订单详情",
        "tags": [
          "订单"
        ]
      }
    },
...
    "/api/admin/products": {
      "get": {
        "operationId": "AdminProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "关键词搜索",
            "schema": {
              "example": "手机",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分类ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "最低价格",
            "schema": {
              "minimum": 0,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "最高价格",
            "schema": {
              "minimum": 0,
              "example": 1000,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string",
              "enum": [
                "price",
                "createdAt",
                "name"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "商品列表（管理后台，含已下架）",
        "tags": [
          "后台管理 - 商品"
        ]
      },
      "post": {
        "operationId": "AdminProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "新增商品",
        "tags": [
          "后台管理 - 商品"
        ]
      }
    },
    "/api/admin/products/{id}": {
      "put": {
        "operationId": "AdminProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "编辑商品",
        "tags": [
          "后台管理 - 商品"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "下架商品（软删除）",
        "tags": [
          "后台管理 - 商品"
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "operationId": "AdminOrdersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "confirmed",
                "shipped",
                "completed",
                "cancelled"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "所有订单列表",
        "tags": [
          "后台管理 - 订单"
        ]
      }
    },
    "/api/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "订单详情",
        "tags": [
          "后台管理 - 订单"
        ]
      }
    },
    "/api/admin/orders/{id}/status": {
      "put": {
        "operationId": "AdminOrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "修改订单状态",
        "tags": [
          "后台管理 - 订单"
        ]
      }
    },
    "/api/admin/orders/{id}/pay": {
      "put": {
        "operationId": "AdminOrdersController_adminPay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "模拟支付（管理员）",
        "tags": [
          "后台管理 - 订单"
        ]
      }
    },
    "/api/admin/categories": {
      "get": {
        "operationId": "AdminCategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "分类管理列表",
        "tags": [
          "后台管理 - 分类"
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "新增分类",
        "tags": [
          "后台管理 - 分类"
        ]
      }
    },
    "/api/admin/categories/{id}": {
      "put": {
        "operationId": "AdminCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "编辑分类",
        "tags": [
          "后台管理 - 分类"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "删除分类",
        "tags": [
          "后台管理 - 分类"
        ]
      }
    }
  },
  "info": {
    "title": "电商 API",
    "description": "NestJS 电商系统 REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "role": {
            "enum": [
              "customer",
              "admin"
            ],
            "type": "string",
            "description": "角色"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "passwordHash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "passwordHash"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "分类ID"
          },
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "description": {
            "type": "string",
            "description": "分类描述"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "parent": {
            "$ref": "#/components/schemas/Category"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "parent",
          "children"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "商品ID"
          },
          "name": {
            "type": "string",
            "description": "商品名称"
          },
          "description": {
            "type": "string",
            "description": "商品描述"
          },
          "price": {
            "type": "number",
            "description": "价格"
          },
          "stock": {
            "type": "number",
            "description": "库存"
          },
          "imageUrl": {
            "type": "string",
            "description": "主图URL"
          },
          "categoryId": {
            "type": "number",
            "description": "所属分类ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否上架"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "stock",
          "categoryId",
          "isActive",
          "createdAt",
          "updatedAt",
          "category"
        ]
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "订单ID"
          },
          "orderNo": {
            "type": "string",
            "description": "订单号"
          },
          "totalAmount": {
            "type": "number",
            "description": "订单总金额"
          },
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "shipped",
              "completed",
              "cancelled"
            ],
            "type": "string",
            "description": "订单状态"
          },
          "paymentStatus": {
            "enum": [
              "unpaid",
              "paid"
            ],
            "type": "string",
            "description": "支付状态"
          },
          "shippingAddress": {
            "type": "string",
            "description": "收货地址"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "description": "支付时间"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "userId": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        },
        "required": [
          "id",
          "orderNo",
          "totalAmount",
          "status",
          "paymentStatus",
          "shippingAddress",
          "createdAt",
          "updatedAt",
          "userId",
          "user",
          "items"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "订单明细ID"
          },
          "productName": {
            "type": "string",
            "description": "商品名称（快照）"
          },
          "unitPrice": {
            "type": "number",
            "description": "单价（快照）"
          },
          "quantity": {
            "type": "number",
            "description": "数量"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "orderId": {
            "type": "number"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "productId": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": [
          "id",
          "productName",
          "unitPrice",
          "quantity",
          "createdAt",
          "orderId",
          "order",
          "productId",
          "product"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "description": "商品名称",
            "example": "无线鼠标"
          },
          "description": {
            "type": "string",
            "description": "商品描述",
            "example": "静音无线鼠标，2.4G连接，续航持久"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "价格",
            "example": 99.99
          },
          "stock": {
            "type": "number",
            "minimum": 0,
            "description": "库存",
            "default": 0,
            "example": 100
          },
          "imageUrl": {
            "type": "string",
            "description": "主图URL",
            "example": "https://example.com/image.jpg"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID",
            "example": 1
          }
        },
        "required": [
          "name",
          "price",
          "categoryId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "description": "商品名称",
            "example": "无线鼠标"
          },
          "description": {
            "type": "string",
            "description": "商品描述",
            "example": "静音无线鼠标，2.4G连接，续航持久"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "description": "价格",
            "example": 99.99
          },
          "stock": {
            "type": "number",
            "minimum": 0,
            "description": "库存",
            "default": 0,
            "example": 100
          },
          "imageUrl": {
            "type": "string",
            "description": "主图URL",
            "example": "https://example.com/image.jpg"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID",
            "example": 1
          }
        }
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "shipped",
              "completed",
              "cancelled"
            ],
            "type": "string",
            "description": "订单状态",
            "example": "confirmed"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "description": "分类名称",
            "example": "电子产品"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "分类描述",
            "example": "手机、电脑、数码配件"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID",
            "example": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "description": "分类名称",
            "example": "电子产品"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "分类描述",
            "example": "手机、电脑、数码配件"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID",
            "example": 1
          }
        }
      }
    }
  }
}